Refactored some code and added retrieve_ldap_node function.
Moved the code that loads an LdapNode's attributes lazily
out to a function to avoid it being repeated in multiple places.

This also allows us to request the attributes to be retrieved
manually, instead of having to access something to force
lazy-loading to work.

Because of all this, a convenience function was added to LdapConnection
which allows us to fetch an object (with all its attributes) from the
database or get an ldap.NO_SUCH_OBJECT exception immediately.